+++ docs/gitbook/python/changelog.md
@@ -2,12 +2,13 @@
 
 <!--next-version-placeholder-->
 
+## v1.23.0 (2023-12-18)
+### Feature
+* **queue:** Add getRateLimitTtl method [python] ([#2340](https://github.com/taskforcesh/bullmq/issues/2340)) ([`f0a1f70`](https://github.com/taskforcesh/bullmq/commit/f0a1f7084478f7899233021fbb4d4307c94dfead))
+
 ## v1.22.0 (2023-12-14)
 ### Feature
 * **job:** Add isFailed method [python] ([#2333](https://github.com/taskforcesh/bullmq/issues/2333)) ([`19bfccc`](https://github.com/taskforcesh/bullmq/commit/19bfccc2d7734b150a5fbb6ea720fcd9887c9dd3))
-
-### Fix
-* **flows:** Update constructor and methods to match queue base ([#2324](https://github.com/taskforcesh/bullmq/issues/2324)) ([`d6c2064`](https://github.com/taskforcesh/bullmq/commit/d6c2064b1fdd88bd4cc61e049ce055ff620b0062))
 
 ## v1.21.0 (2023-12-14)
 ### Feature